Adding routes and db management
