+++ src/util/db.js
@@ -1,0 +1,16 @@
+const config = require('./../config/config');
+const mysql = require('mysql'); 
+
+const connection = mysql.createConnection({
+    host: config.database.host,
+    user: config.database.user,
+    password: config.database.password,
+    database: config.database.dbname
+});
+
+connection.connect(function(err) {
+    if (err) throw err;
+    console.log("Connected!");
+});
+
+module.exports = connection;
